refactor: move files in /store to activate Vuex modules
